Move language into common lib
