+++ src/Common/Common.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+    <Nullable>annotations</Nullable>
     <RootNamespace>InstantTraceViewer.Common</RootNamespace>
   </PropertyGroup>
