+++ src/InstantTraceViewerUI/InstantTraceViewerUI.csproj
@@ -10,7 +10,7 @@
     <SupportedOSPlatformVersion>10.0.22621.0</SupportedOSPlatformVersion>
     <StartupObject>InstantTraceViewerUI.Program</StartupObject>
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <Nullable>disable</Nullable>
+    <Nullable>annotations</Nullable>
     <UseWindowsForms>True</UseWindowsForms>
     <ApplicationIcon>Assets\Logo.ico</ApplicationIcon>
     <PublishReadyToRun>true</PublishReadyToRun>
